Don't send meta events to the external instrument
While the MacOS backend allows this, and the song still works, it causes
a "non-sysex message must not be longer than 3 byte" error on the
Windows backend. Skipping all meta events avoids this problem, and is
probably more efficient anyways. I don't think meta events are used when
sent to an external instrument to play in real time; rather, they may only
be used in things like DAWs.